Simplify API requirements for quantifiers.
By assuming that the assemble_quantification_data.py script is going to be
executed in the directory in which quantification took place, we can remove the
method get_results_file() and read_transcript_abundances() - i.e. the method
get_transcript_abundance() should know, relative to the quantification
directory, where to find the file containing transcript abundances.