PYTHONPATH doesn't need to be adjusted on new install.
The correct way to package a Python app is somewhat inscrutable to me.
Hopefully this is correct now - it means that the piquant package should
automatically be on the Python path, and the scripts should be on the
executable path within the virtual environment in which piquant has been
installed. An upside to this is that the location of the bin directory doesn't
need to be passed around so that it can be used in prepare_read_simulation.py
and prepare_quantification_run.py.

This checkin also fixes tests which didn't work in Python 3.