Improve reliability of between-run comparisons.
Because most statistics were reported for the set of transcripts (or genes) for
which both real and calculated TPMs were greater than a threshold value, there
has always been an issue with comparisons between runs not being truly like
with like - differences in the behaviour of different quantification tools
would lead to different "true positive" sets in each case.

Instead, I have taken a leaf from this comparison of Kallisto and Salmon by Rob
Patro:

http://nbviewer.ipython.org/gist/rob-p/4e1285ac94456a46e3ce

Now, any real or calculated TPMs that are below a user-defined threshold are
set to zero. This has two effects:

i) Most statistical comparisons are carried out on the set of transcripts (or
genes) with non-zero real TPMs after truncated - this means that the same set
of transcripts is considered for any run of a quantification tool against a
particular set of simulated data.

ii) Quantification tools which report a small non-zero TPM for transcripts (or
genes) which have zero real TPM are not punished unnecessarily for doing so,
when such a difference between calculated and real TPM is likely of little
biological interest.