Update indexing and quantification calls for Sailfish
The interface of Sailfish (>= v0.7.0) in terms of both
command line parameters and output format has changed
slightly (to be concordant with Salmon).  Additionally,
the newer version of Sailfish are faster and more
accurate.  This commit updates the code to interact
with Sailfish via it's new interface (removing the
unnecessary -k parameter from indexing, modifying the
library format specification, and altering the way
the output is read to properly extract the TPMs).