Replace TopHat with STAR for Cufflinks quantification.
Also, revert previous change (sorting mapped reads) for Express. This was done
because updating Bowtie from 1.0.0 to 1.2 resulted in mappings for reads no
longer being written consecutively (a problem for Express). But then sorting
the reads by name caused a problem because of the ordering inherent in Flux
Simulator read names (i.e. it results in them being ordered by chromosome). So
have reverted to Bowtie 1.0.0.