Propitiate the angry gods of Python packaging.
For some reason matplotlib will no longer install correctly via pip - it
appears to download and install numpy itself in order to have this to build
against, but the version it downloads is not the correct version that it later
seems to require for installation (??). Have added an installation script to
ensure that the version of numpy that we desire is installed prior to
matplotlib being installed, since this ordering does not appear to be
achievable through setup.py itself.